Fixes formating of 2 strings Should fixed assert errors [1385707]

+++ src/hugin1/base_wx/huginConfig.cpp
@@ -74,7 +74,7 @@ void FillPlaceholders(Placeholdersmap & placeholder, const HuginBase::Panorama &
 
     wxFileName lastImg(wxString(pano.getImage(pano.getNrOfImages()-1).getFilename().c_str(),HUGIN_CONV_FILENAME));
     placeholder[wxT("%lastimage")]=lastImg.GetName();
-    placeholder[wxT("%#images")]=wxString::Format(wxT("%d"), pano.getNrOfImages());
+    placeholder[wxT("%#images")]=wxString::Format(wxT("%lu"), (unsigned long)pano.getNrOfImages());
     PanoramaOptions opts=pano.getOptions();
     pano_projection_features proj;
     if (panoProjectionFeaturesQuery(opts.getProjection(), &proj))
